requirements.txt
aniso8601==3.0.2
certifi==2018.4.16
chardet==3.0.4
click==6.7
cycler==0.10.0
decorator==4.3.0
elasticsearch==6.3.0
Flask==0.12.2
Flask-Cors==2.1.0
Flask-RESTful==0.3.4
Flask-Session==0.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
future==0.16.0
geocoder==1.38.1
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
joblib==0.12.2
kiwisolver==1.0.1
MarkupSafe==1.0
numpy==1.15.0
pandas==0.23.4
pipenv==11.10.0
psycopg2==2.7.4
pyspark==2.2.1
python-dateutil==2.7.3
pytz==2018.5
ratelim==0.1.6
requests==2.18.4
scikit-learn==0.19.1
scikit-surprise==1.0.6
scipy==1.1.0
six==1.11.0
SQLAlchemy==1.2.6
urllib3==1.22
virtualenv==15.2.0
virtualenv-clone==0.3.0
Werkzeug==0.14.1
WTForms==2.1
aniso8601==3.0.2
certifi==2018.4.16
chardet==3.0.4
click==6.7
cycler==0.10.0
decorator==4.3.0
elasticsearch==6.3.0
Flask==0.12.2
Flask-Cors==2.1.0
Flask-RESTful==0.3.4
Flask-Session==0.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
future==0.16.0
geocoder==1.38.1
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
joblib==0.12.2
kiwisolver==1.0.1
MarkupSafe==1.0
numpy==1.15.0
pandas==0.23.4
pipenv==11.10.0
psycopg2==2.7.4
python-dateutil==2.7.3
pytz==2018.5
ratelim==0.1.6
requests==2.18.4
scikit-learn==0.19.1
scikit-surprise==1.0.6
scipy==1.1.0
six==1.11.0
SQLAlchemy==1.2.6
urllib3==1.22
virtualenv==15.2.0
virtualenv-clone==0.3.0
Werkzeug==0.14.1
WTForms==2.1
